requirements.txt
selenium 
webdriver-manager 
pandas
python-dotenv
streamlit
langchain
langchain-google-genai
nltk 
vaderSentiment 
chromedriver-autoinstaller